requirements.txt
pandas
matplotlib
seaborn
scipy
pillow
wordcloud
tweepy
lxml
pyquery

